s3-security/requirements.txt
attrs==24.2.0
aws-cdk-lib==2.170.0
aws-cdk.asset-awscli-v1==2.2.212
aws-cdk.asset-kubectl-v20==2.1.3
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==38.0.1
boto3==1.35.68
botocore==1.35.68
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
constructs==10.4.2
dxrpy==0.2.12
idna==3.10
importlib_resources==6.4.5
jmespath==1.0.1
jsii==1.105.0
publication==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pulumi>=3.0.0,<4.0.0
pulumi-aws>=5.0.0,<6.0.0
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
